Refactor behaviour triggered by a flag into separate classes.
